feat: Improve board identification logic in MPRemoteBoard fixes: https://github.com/Josverl/micropython-stubber/issues/558
The code changes in this commit improve the board identification logic in the `MPRemoteBoard` class. Specifically, the `find_board_id_by_description` function now takes an additional `version` parameter to filter the board information based on the firmware version. This allows for more accurate board identification.

This commit also updates the default value for the `self.board` attribute in case the board cannot be identified. Instead of setting it to "UNKNOWN", it is now set to "UNKNOWN_BOARD" for clarity.

The changes in this commit address the need for better board identification in the `MPRemoteBoard` class, improving the overall functionality of the code.